Fixed Fragment Implementation
