+++ app/src/main/java/com/example/socketexperiment0/ui/receive/ReceiveViewModel.kt
@@ -6,10 +6,13 @@ import androidx.lifecycle.ViewModel
 
 class ReceiveViewModel : ViewModel() {
 
+    /*
     private val _text = MutableLiveData<String>().apply {
         value = "This is dashboard Fragment"
     }
-    //val text: LiveData<String> = _text
+    val text: LiveData<String> = _text
+
+     */
 
 
 }
