+++ app/src/main/java/com/example/socketexperiment0/ui/send/SendViewModel.kt
@@ -1,13 +1,19 @@
 package com.example.socketexperiment0.ui.send
 
-import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import com.example.socketexperiment0.R
+import com.google.android.material.textfield.TextInputLayout
 
 class SendViewModel : ViewModel() {
 
+    /*
     private val _text = MutableLiveData<String>().apply {
         value = "This is Send Fragment"
     }
     val text: LiveData<String> = _text
+
+     */
+
+
 }
